# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Faz a lógica por trás do sistema FIFO"
  version: "1.0.0"
  title: "FIFO Api"
  termsOfService: "http://localhost:3333/api/users/termos"
  contact:
    name: "Suporte"
    url: "http://localhost:3333/api/users/suporte"
    email: "souza.joaoguilherme2002@gmail.com"
  license:
    name: "MIT"
    url: "http://opensource.org/licenses/MIT"
host: "206.189.189.136:3333"
...
schemes:
- "http"
- "https"
paths:
...
  /api/queues:
    get:
      tags:
      - "queues"
      summary: "Obtem os dados da plataforma"
      description: "Essa rota retorna os dados das plataformas"
      operationId: "getAllQueueData"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Returned with sucessfully"
          schema:
            $ref: "#/definitions/inline_response_200_8"
        "500":
          description: "Server erro"
          schema:
            $ref: "#/definitions/inline_response_500_4"
      x-swagger-router-controller: "Queues"
    post:
      tags:
      - "queues"
      summary: "Insere o usuário na fila"
      description: "Essa rota cria o registro ou atualiza o status na fila"
      operationId: "storeQueue"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "platform"
        required: true
        schema:
          $ref: "#/definitions/platform"
      responses:
        "201":
          description: "User is on queue"
          schema:
            $ref: "#/definitions/inline_response_201_4"
        "500":
          description: "Server error"
          schema:
            $ref: "#/definitions/inline_response_500_5"
      x-swagger-router-controller: "Queues"
  /api/queues/exit:
    put:
      tags:
      - "queues"
      summary: "Remove o usuário da fila"
      description: "Essa rota atualiza o status na fila para 0"
      operationId: "removeQueue"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "platform"
        required: true
        schema:
          $ref: "#/definitions/platform_1"
      responses:
        "201":
          description: "User is off the queue"
          schema:
            $ref: "#/definitions/inline_response_201_5"
        "500":
          description: "User or platform not found"
          schema:
            $ref: "#/definitions/inline_response_500_6"
      x-swagger-router-controller: "Queues"
  /api/queues/polling/{platform}:
    get:
      tags:
      - "queues"
      summary: "Obtem a posicao do usuario"
      description: "Essa rota retorna os dados das plataformas"
      operationId: "polling"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "platform"
        in: "path"
        description: "ID of platform to get the user position"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Success"
          schema:
            $ref: "#/definitions/inline_response_200_9"
        "500":
          description: "server error"
          schema:
            $ref: "#/definitions/inline_response_500"
      x-swagger-router-controller: "Queues"
...
definitions:
...
  inline_response_500_6:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "Server error, try again!"
  inline_response_500_5:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "Server error."
  inline_response_500_4:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "Error on request."
...
  platform:
    type: "object"
    properties:
      platform:
        type: "integer"
        example: 1
...
  inline_response_500:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "Error on server"
...
  platform_1:
    type: "object"
    properties:
      platform:
        type: "integer"
        example: 1
  inline_response_200_9:
    type: "object"
    properties:
      position:
        type: "integer"
        example: 6
    example:
      position: 6
...
  inline_response_201_4:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "User is on queue"
      position:
        type: "integer"
        example: 2
      notification:
        type: "boolean"
        example: true
    example:
      msg: "User is on queue"
      notification: true
      position: 2
...
  inline_response_201_5:
    type: "object"
    properties:
      msg:
        type: "string"
        example: "Exited with sucessfully"
    example:
      msg: "Exited with sucessfully"
...
  inline_response_200_8:
    type: "object"
    properties:
      id_platform:
        type: "integer"
        example: 2
      size:
        type: "integer"
        example: 11
    example:
      size: 11
      id_platform: 2
...